+++ tests/CMakeLists.txt
@@ -52,7 +52,6 @@ set(COMMON_SOURCES
     ${CMAKE_CURRENT_SOURCE_DIR}/ComTests.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/FileSystemTests.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/NTResultTests.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/PropertyTests.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/ResourceTests.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/ResultTests.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/Rpc.cpp
